linux: skip clone if URL is set by CLI
We support two ways to clone a repo:

  a) Host side
  b) guest side

The CLI definition of a git tree URL recently added via commit a403bd6fb
("linux: add support for setting linux URL and ref from CLI") allowed us
to customize the tree and ref to use. The assumption was that we would
skip the clone done by kdevops itself for the defined Linux tree, but in
practice it actually tries again. So let's be explicit and skip the
clone on the hose side only if the URL was set through the CLI.

While at it, let's be explicit about requirements and require the
directory to be created before we proceed.

Fixes: a403bd6fb1845a36 ("linux: add support for setting linux URL and ref from CLI")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>