guestfs: make debian trixie default
kdevops relies on rolling distributions. We've been stuck with
debian12 because that's what the virt-builder servers have and
it's a bit old for our taste in kernel hacking the latest and
greatest kernels.

Since we are now boosting efforts on CI kernel hacking testing,
we want to make bringup as fast as possible, and since the first
thing we typically want is the *latest and greatest* we tend to
sometimes upgrade a system on bringup. Even if we skip that,
buster 12 default is still a bit... old... and so to help speed
up bringup in both cases where we just want the latest debian
testing *or* an updated system, make debian 13 trixie the default
image now that we've fixed it.

Tested with:

make linux
make fstests

And we've been running tests on guests where the host is trixie and the
guest is debian 12 (buster) but then due to the host having a local apt
mirror, we rely on CONFIG_GUESTFS_COPY_SOURCES_FROM_HOST_TO_GUEST to
do immediately do upgrades from debian 12 --> debian 13. And debian 13
has been tested for many of our workflows including the latest sysbench
workflow.

So all in all, debian trixie has been being tested for a while, it was
just this painful upgrade process we've had to deal with for a long
time.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>