guestfs: fix debian13 nsswitch.conf edit
The debian13 guests come with an nsswitch.conf line like this:

hosts:          files myhostname resolve [!UNAVAIL=return]

When we have CONFIG_GUESTFS_HAS_DISTRO_SOURCE_COPY_CAPABILITIES=y
which would be when we have say a debian13 host and a debian13 guest
and the host has a hop1 /etc/apt/sources.list, we copy the host's
sources.list so we can profit from the speedy mirrors you have.

However we need DNS to make sure we can talk to the servers and the
above line prevents that. I had first tried to just remove the UNAVAIL
entry with virt-builder command:

edit /etc/nsswitch.conf:'s/\[!UNAVAIL=return\]//'

But for the life of me this is not taking effect. So just do a brute
force wipe of the file and open code the entire set of lines we want
for debian13 for /etc/nsswitch.conf.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>