guestfs: enhance GUESTFS_COPY_SOURCES_FROM_HOST_TO_GUEST
When we detect that your host has hop1 to its sources on debian,
and your guest can also leverage this we ensure we copy the host's
local /etc/apt/sources.list automatically. This is critical for
self-hosted runners as it means reducing the amount of time it takes
to update a system or install packages.

An issue keeps coming up despite ongoing mirror best practices is
that the release file is signed but not valid for a period of time:

Failed to update apt cache: E:Release file for
http://some-custom-debian-mirror.local:8081/artifactory/master/dists/testing/InRelease
is not valid yet (invalid for another 4h 33min 45s). Updates for this
repository will not be applied., E:Release file for
http://some-custom-debian-mirror.local:8081/artifactory/master/dists/testing-updates/InRelease
is not valid yet (invalid for another 4h 33min 45s). Updates for this
repository will not be applied.

I've checked NTP is set on the host and the time matches with the end
mirror, and so this is just an odd timing issue. We recently tried to
address this through commit 13239715371c0 ("devconfig: add chicken-and
egg debian mirror workaround") but this is done just for the first
devconfig update / upgrade -- when devconfig_try_upgrade is True. But
this is not the only time we're going to try to apt-get update. So
instead of making this a on-off thing, make this permanent when we use
the heuristics in guestfs to copy the sources to the guest, that is
guestfs_copy_sources_from_host_to_guest == True. Right now this is only
being leveraged by debian hosts / target node systems only and so this
does not affect other run times.

This fixes stalls on refreshing sources.list on package installs on
self-hosted runners with this setup.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>