guestfs: move kdevops_enable_guestfs definition to Kconfig
Instead of relying on Makfile magic use output yaml. We could
have usted the bool GUESTFS, but sicne the explicit language of
"kdevops_enable_guestfs" was used before and is used elsewhere,
just keep this for now. We can later replace this but to ensure
we don't break old runtime environments which may already be
configured and do not have guestfs as a bool we need to keep
backward compatibility.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>