bringup_guestfs: add dnsmsasq check for debian based distros
I can vet that at least on debian based distros we have a conflict when
you have dnsmasq installed and want to leverage the guestfs default
network bringup. You typically end up having to work around this by
disabling the dnsmasq service first. I've checked many systems I have
kdevops on and none of them have dnsmasq. For now use this only on
debian based systems, we can propagate this to other distros when we verify
this is also a sane check for them.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>