distro: use distro_debian_has_hop1_sources for the debian release time workaround
The GUESTFS_COPY_SOURCES_FROM_HOST_TO_GUEST timing issues we are seeing
are currenlty observed on debian testing systems which have hop1
mirrors, but we're enabling this timing work around even when you don't
have hop1 mirrors.

Relax the heuristic to only be used on hop1 mirrors and when we copy
over the sources from the host. We can later relax this even when we
just copy over the sources in general to the guest but for now I don't
have full proof this is needed. I think this is just a minor mirror
misconfiguration of some sorts.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>