host: propagate distro vars
Propagate distro variables in case we need to do specific distro
checks, or we have only tested some ansible tasks / heuristics on
some distros.

While at it, add a heuristic bool which can be used to be able to
generalize which base distribution your distribution is based on
since we can often can use or checks for distributions based on the
same family. For example debian based and ubuntu based distros are
debian based, and issues found on debian likely are found on ubuntu.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>