devconfig: enhance debian enhance distro_debian_has_hop1_sources heuristic
When the host has debian, the guest has debian and both match release
target or we know there is a way to move the guest forward, we use
a heuristic to check if /etc/apt/sources.list on the host has a hop of
1 to it, we default to allow the guest to get the host's sources.list
file.

Debian 13 Trixie however has this on /etc/nsswitch.conf:

files myhostname resolve [!UNAVAIL=return] dns

And the [!UNAVAIL=return] dns will preven the guest from using DNS
for the apt mirror for some reason.

Although commit a400440e853b6 ("guestfs: fix debian13 nsswitch.conf
edit") fixed this when we download a fresh debian guestfs image, we
should not assume the user has the latest image in their cache. Instead
just ensure we check for this and fix it.

This enhance our CI. While at it, also enhance our looseness on the
mirror requirements. This will avoid more odd timing issues on the
guests.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>