journal-server: add support for 'make journal-dump'
This will let us run this for CIs so to leverage not only the
guestfs console log but another flavor in ASCII form of the systemd
journal which may be more complete and easier to read.

All we need to do now is run 'make journal-dump'

We support running this 'make journal-dump' even if you have
CONFIG_DEVCONFIG_ENABLE_SYSTEMD_JOURNAL_REMOTE disabled as its
a nice future source of kernel logs. Even if you run on the cloud
it should be possible to set this up with a central command
and control base. But we don't error our as we want all CIs to
leverage this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>