defconfigs: enable systemd remote journal on CI selftests
We want to enable CONFIG_DEVCONFIG_ENABLE_SYSTEMD_JOURNAL_REMOTE=y
on selftests as this can let us capture the kernel journals from
systemd, and gives us a second mechanism to capture the kernel log
from the target nodes.

We enable this on the cli selftests, that is, the selftests used
where we specify the LINUX_TREE and LINUX_TREE_REF on the command line
interface, used for automated CI tests.

We don't enable this on the non-cli defconfigs as folks may want to
run these test on the cloud, and there is no setup yet for managing
CONFIG_DEVCONFIG_ENABLE_SYSTEMD_JOURNAL_REMOTE=y on the cloud. That
can easily be done by creating an extra system and using that as the
source for where to stash the remote journals.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>