CI: add just one fstests test for the kdevops fstests test
We don't want to force testing a full filesystem test on kdevops
for every patch / push, but we want to verify that at least one
test can run, and that test collection works is no one breaks it.
We can address this by just limiting the scope of tests by using:

make fstests-baseline TESTS=generic/003

Which will just run the generic/003 test.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>