CI: reduce scope and use github repo private secrets
The ssh-agent is not used or leveraged by github actions, and so we
in order to automatically push onto kdevops-results-archive you will
need to leverage github repository secrets for the key and the
passphrase.

Let's reduce the scope as we test this so we can test these simple
changes faster.

Use the demo repository for now.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>