CI: fix build and processing results
The expectation was that we'd need an ssh key with a passphrase.
The github way to leverage an ssh-agent is through webfactory [0 and
they insist that since the key is uploaded to github that suffices. The
passphrase is not supported so remove it.

The local user ssh-agent is not used or leveraged by github actions,
and so we in order to automatically push onto kdevops-results-archive
you will need to leverage github repository secrets for the key, once
that is installed you will also need to add the public key to the
kdevops-repository-results as a "deploy" key.

 - fstests.yml:
   - use webfactory/ssh-agent v0.9.0
   - simplify the way we look for zip files on artifacts, just use a glob
   - fix the fstests.yml kdevops path for the zip file
 - kdevops_archive:
    - reduce scope and use github repo private secrets
    - add debug info about archive files

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>