kdevops_archive: enhance commit and archive placement
The commit log on kdevops-results-archive does not provide much
details, we can't even tell what tree this was triggered from.
So enhance all this and add extra CI files the test runner could
populate in case it wanted to help users of kdevops-results-archive
or bots that scrape it for analysis of results.

New optional CI metadata files for kdevops archive:

  ci.trigger   - should be set by the repo basename name that triggered the test
                 like linux-radix-tree-kpd or linux-modules-kpd or
                 kdevops ... and so on
  ci.subject   - should be set to the last commit in your repo
                 tree that triggered this CI event, so this was designed
                 for something like 'git log -1 --pretty=format:"%s"'
  ci.result    - Test outcome status:
		 - ok
                 - not ok
                 - WTF
  ci.commit_extra - Additional details to include in commit message into
                    kdevops-results-archive. Whatever you want buddy.

All files are optional with these defaults:
  ci.trigger   -> target_linux_tree
  ci.subject   -> "${target_linux_tree} ${target_linux_ref} result #${test_number}"
  ci.result    -> "unknown"
  ci.commit_extra -> "No extra test report details available."

Also use the subject prefix of the ci.subject as you would get if
you use git format-patch on it, for the directory name where we
place your zip file. This lets you easily find the zip file for
posted patches on the mailing list fast.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>