devconfig: retry removal of unattended-upgrades
Experience is showing we can run into a race with unattended-upgrades
already kicking off... so retry that removal hard. This will retry
up to 20 times every 30 seconds, for a total of up to 10 minutes,
otherwise we fail. If this still fails we'll have to work to remove
this on the image with virt-builder itself.

The affected guests are debian with this patch.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>