kdevops_archive: force local mirror to account for epoch changes
Part of the strategy of kdevops-results-archive is that it should be
able to grow up to the max limit size of the repo allowed, and then
when its reached capacity all we have to do is rename it to associate
it with an epoch like kdevops-results-archive-2024-10 and we then
create a new tree to start fresh. When this rotation happens the mirrors
will actually pick this up automatically and old data is just ignored.

However we must ensure that github runners don't keep local copies
around of the repo, other than the mirror. So adjust the mirror settings
to phone home to 10s and also just remove the runner's copy of the
kdevops-results-archive every time, and clone a fresh copy by requiring
a local git copy.

The kdevops-results-archive-2024-10 was created as a test but in this
case we didn't start fresh, we just kept all the old tarballs as we
haven't yet reached capacity. We simply did the epoch change to try
to force and test for when this will happen eventually and to make
the transition later easier.

If you don't have a local mirror for the kdevops-results-archive we
just gracefully bail out. This allows for github runners to still
leverage and use the zip file created for local artifacts. The only
thing that would be skipped would be the commit of the respective
archive xz file into kdevops-results-archive with a nice commit log
message describing the results.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>