workflows/linux/refs: use similar naming for modules next ref
The XFS tree already uses the postfix REF_NEXT for the next ref,
but the modules tree has it a NEXT_REF. We also had not run

make refs-default

in a while, so run it and this updates our files in the directory
workflows/linux/refs/default/ for us based on the yaml input files.

No defconfig uses these yet so we're good to go with this simple
rename right now. That is there are no affected users.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>