defconfigs: remove the need for CLI specific defconfigs
The need for the CLI defconfigs was due to the fact we had to disable
wanting to disable Linus' tree when we use a definition of a
kernel tree and ref through the CLI.

But its silly to require cli defconfigs if we can just disable
Linus' tree when the CLI is used to define a tree.

We can accomplish this by just forcing a custom tree when the
CLI is used and making also Linus' tree the default otherwise.
This later step makes it a bit clearer that Linus' tree is
preferred for much subsystem testing.

We can addres defconfigs which *want* custom trees other than Linus'
tree later and which want to also allow to override the tree through
the CLI, the difficulty there will be that these defconfig *want* to
*opt-in* to a specific tree on the defconfig. The current solution
is that we make Linus' tree *default*, and this makes it easier for
us to just specify custom trees on the command line. Likely the
solution is to just have custom trees be part of the CLI like for
example USE_MODULES_TREE or something like that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>