selftests: create /lib/firmware for firmware selftests
The first test which the firmware selftests runs assumes you
have /lib/firmware directory created as we want to mount /lib/firmware
onto tmpfs. This test was added via upstream Linux commit
901cff7cb9614 ("firmware_loader: load files from the mount namespace of
init"). The test will fail if the directory does not exist. We'll
send a fix for that, but in the meantime, ensure that kdevops
creates the directory for us on the target test node.

The test was failing because the directory does not exist, and it
gives the impression the test fails, but this is false positive,
the test needs improvement so to ensure the directory exists
before it tries to mount. I'll send a fix after this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>