guestfs: enhance ssh config choice
The ssh configuration file we use when we set up guests with guestfs
is is simply ~/.ssh/config_kdevops_$KDEVOPS_HOSTS_PREFIX. This has
scaled for a while but now that we have CI support each self-hosted
will want to use the relevant git commit ID upstream from the upstream
repo as the prefix, when that is used then technically we can't test
the same test for different subsystems. Additionally even we can't
run kdevops on the same host with the default prefix on the same distro
more than once without creating an overwrite of the file. For example
debian13 systems will use .ssh/config_kdevops_debian13 for both XFS
and even if NFS tests.

Enhance support to scale by having a unique prefix for the ssh
configuration. We choose the current working directory for kdevops
as the unique pivot for what we'll use, so we can just compute the
sha256sum of the current working directory.

To remain backward compatible add TOPDIR_PATH_HAS_SHA256SUM which
default to y, so that in older setups if folks git pull, things will
still work with the old prefix.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>