fstests/Kconfig: use int-specific default for CLI SOAK_DURATION
Commit e6294385d2b3b6 ("fstests: add support for overriding soak
duration through CLI") added support for overriding the soak duration
through the CLI which is useful for CIs. It however re-used an existing
script we use for Kconfig defaults, this works well for strings but
the default output must be digit if the Kconfig variable was an int,
so just add another script for int Kconfig variables so we can
avoid the warning, the warning from Kconfig is that we essentially
default to "" if the CLI option to override the soak duration was
was not set, and this needs to be 0.

Fixes: e6294385d2b3b6 ("fstests: add support for overriding soak duration through CLI")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>