fstests: add support to customize groups through CLI
It would be good to allow CIs to customize the group we want to test.
Right now we default to "auto" but it would be good for a CI to let
the user drop down a menu and click on "quick" for instance.

Demo:

$ make mrproper
$ make defconfig-xfs_reflink_4k FSTESTS_GROUP="quick"
$ grep FSTESTS_RUN_CUSTOM_GROUP_TESTS .config
CONFIG_FSTESTS_RUN_CUSTOM_GROUP_TESTS="quick"

Suggested-by: Carlos Maiolino <cem@kernel.org>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>