defconfigs: add all XFS profiles
Add a defconfig for all supported XFS profiles. This will allow us
to leverage selecting a test to run on only one specific defconfig
which is specific to a set of XFS profiles. We also provide bundles
for LBS. So we have xfs_nocrc_lbs xfs_reflink_lbs (already merged).

This will allow us to leverage this now to enable our CI infrastructure
to let developers reduce the scope of which profiles they may want to
test.

We allow users then now to pick one out of 30 XFS profiles to test.
The default is to test all.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>