run_kernel_ci.sh: enable guestfs
Support resetting hung systems which our watchdog detects when
guestfs is used. Contrary to vagrant we just use the full name of
the host.

We'd still need a unified way or a way to reset a random cloud node.
Perhaps the way to do that is just make it a cloud specific hook.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>