fstests: enable FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS on CI systems
CI systems will use something like:

make defconfig-xfs LINUX_TREE=https://github.com/linux-kdevops/linux-xfs-kpd.git

When this is used BOOTLINUX_TREE_SET_BY_CLI is y, and so we can
just modify the existing distro heuristic for when a distro prefers
a reset on hung system detection:

  HAVE_DISTRO_PREFERS_FSTESTS_WATCHDOG_RESET

So we default that to y when BOOTLINUX_TREE_SET_BY_CLI is set. So
with the above command we'd end up with:

grep FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS .config
CONFIG_FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS=y

This will allow us to force a reset on a target node when libvirt
is used, for guestfs we'll try sudo virsh reset. The default
timeout is 60 minutes for a new test. Since each CI uses a "new"
test, then that's our limit. We can later enhance this by leveraging
prior art check.time values but we'd need to also match the soak
duration configuration, so a bit more thought is required for this.
For now this should suffice to at least test trying to see if we can
pick up kernel crashes and if we reset a system accordingly.

Post-mortem  analysis of a crash to highlight a crash happened should
be evaluated next, because as-is we'd just reset the system and kdevops
would go on with the test, and its not clear if xunit picks up on the
failed test.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>