.ci/results/fstests: add .kernel.ci.* files
We already enable kernel-ci on kdevops, but reduced
the steady state to 1 on commit 33bd44f947011c ("kerenl-ci:
modernize docs") to make effective communication over the value
of the steady state in light of kdevops' adoption of soak duration
and support for it.

But since kdevops-ci has was always been enabled all it does is
expose our 'make fstests-baseline-loop' target, with a steady
state of 1, it just means the watchdog kicks off and we can
leverage the .kernel.ci.status file for the actual output of
the status of the test DUTs. The CI can just use that, and since
commit ee72aadfd5ba913 ("fstests: enable
FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS on CI systems") enabled the
FSTESTS_WATCHDOG_RESET_HUNG_SYSTEMS, we can now test to see if
we can detect a crash and if xunit processes it correctly.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>