linux: bump us up to next-20241108
Add next-20241108 as the latest linux-next we manually booted and
tested. Since we're bumping up also enable TEST_KALLSYMS=m so we can
leverage that test for defconfigs which will enable it. This will
slightly increase build time but its fine, the current self-hosted
runners are beefy enough and we want this as part of our build
coverage as well anyway.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>