defconfigs/linux-modules-kpd: expand test coverage to include kallsysms
The defconfig for defconfigs/linux-modules-kpd was just a symlink to
the defconfig/selftests-kmod and that just covers *one* test. Expand
on this by allowing us to also include the new kallsyms selftest.

So we add a defconfigs/seltests-modules and allow the
linux-modules-kpd tree to just use that as a symlink, so we
can later just expand on seltests-modules for future modules
testing.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>