devconfig: remove 89-ethernet.network
The work on commit 683610d655563 ("debian: install
/etc/systemd/network/89-ethernet.network") tried to do the best we can
with the willy bug present on systemd-networkd-wait-online.service where
if that service is enabled and you have no configuration for ethernet
the service never completes and fails and this can often cause some
services to time out that we may want to setup.

It turns out that adding our own /etc/systemd/network/89-ethernet.network
just to annotate we need dhcp is not sufficient to fix the issue across
all the self-hosted runners used by kdevops. Although many CIs are
not affected one system, the one used for the linux-firmware CI using
the linux-firwmare-kpd kept failing at bringup. Upon closer inspection
on the console, even if we upgraded the host and rebooted into the
latest debian testing packages, the host still had issues. The console
on the guest revealed that the IP address shifted when the file
89-ethernet.network was installed and the server restarted, so the
timeout on bringup was caused because the IP address we obtained first
for ssh-detection changed, so we can no longer connect to the guest.
Strange enough, if you reboot the guest we see the guest afterwards
gets *two* IP addresses assigned!

This is just nuts. So disable systemd-networkd-wait-online.service as
that is completely buggy and I don't think folks really thought this
out well.

Fixes: 683610d655563 ("debian: install /etc/systemd/network/89-ethernet.network")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>