fstests: Fix NFS mount options for fstests
If the NFS mount command line does not specify an NFS version, the
client uses the default protocol version , which for recent versions
of nfs-utils, is NFSv4.2.

I noticed the NFS test sections in our nfs.config file specifies the
NFS version for the test partition, but not the scratch partition.
This is throwing off the feature test macros in fstests, so it is
testing, say, xattrs, on NFSv4.1 mounts, and failing 100% of the
time.

Ensure that the test and scratch partitions use the same NFS
version.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>