codereadyrepo: Refactor to prepare for more complex usage scenarios
I'm about to add a few more cases that the codereadyrepo role needs
to handle. The only difference for all of these cases is the name of
the repo to enable, so pull out the selection of that name, then do
the repo enablement task only when a repo name was found.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>