bringup: prevent the use of vagrant
Vagrant virtualization has been deprecated for some time, but its
been left in kdevops to allow time for folks to move over to
guestfs, which has been the default for a while.

It's time to start removing Vagrant, so remove the menu choice to
select the use of Vagrant virtualization. This change will be quick
to revert if there are loud complaints.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>