devconfig: install-deps is skipped in some configurations
A recent commit replaced import_tasks with include_tasks in a few
steps in the devconfig playbook.

The Ansible "import_tasks" documentation says:

> Tags are not interpreted for this action, they are applied to the
> imported tasks

Revert some of the import_tasks conversions to ensure those steps
are run according to the command-line tag settings.

For the original problem that I tripped on, add tags to ensure those
sub-steps get run as needed.

Reported-by: Joel Granados <joel.granados@kernel.org>
Closes: https://lore.kernel.org/kdevops/20250128112453.rawxqvpffxwupvco@AALNPWDAGOMEZ1.aal.scsc.local/T/#t
Fixes: ec9a34eb6b8e ("devconfig: Replace import_tasks with include_tasks")
Reviewed-by: Daniel Gomez <da.gomez@samsung.com>
Tested-by: Daniel Gomez <da.gomez@samsung.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>