github/workflows/fstests.yml: use debug callback plugin
Use the new cli ANSIBLE_CFG_CALLBACK_PLUGIN to configure the debug
callback plugin for CI.

This keeps the output same level of verbosity for CI, before we change
the default to dense.

An alternative is to keep it dense but increase the Ansible verbosity
level with AV= makefile parameter.

Signed-off-by: Daniel Gomez <da.gomez@samsung.com>