volume_group: Detect the /data partition directly
The volume_group role currently runs during "bringup", but the
/data partition is set up later by subsequent make targets (eg, when
the test kernel is built or when workflows are set up).

This means the volume_group role has to avoid using one of the block
devices so that roles that run later can put the /data partition on
the skipped device.

Instead of counting the number of available block devices and just
skipping one of them, explicitly avoid using the device listed in
{{ data_device }}. This is a more specific and reliable check.

Works only for guestfs. Support for terraform providers is next.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>