codeready_builder: Fix Oracle Linux 9 CRB
The OL9 CodeReady Builder repo is the same for OCI instances
and local guestfs.

Add an error message to indicate when a code change / improvement
is needed to select the correct CRB repo. It's not possible for us
to get this right in a single commit due to the large number of
possible combinations.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>