ext4: add baseline for v6.14-rc3
So far I've run a full test twice and found a crash which has been
reported [0] to the ext4 mailing list. The crash hits on truncation
and requires CONFIG_DEBUG_VM enabled kernel. I'm disabling this test
on further tests for all ext4 profiles as it can prevent further
evaluation and tests.

Link: https://lkml.kernel.org/r/Z7vIBq-Zuo6Z7ihr@bombadil.infradead.org # 0
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>