terraform/OCI: Add a default VCN
The kdevops set-up for other cloud providers assumes that kdevops
will provision a VCN/subnet for use during test runs. OCI does not
have that option; it assumes that a long-lived subnet already
exists to which kdevops instances should be attached.

In a moment I will introduced a Kconfig option to use network
resources that kdevops manages instead of a pre-existing subnet.
This patch adds those resources, but does not yet use them, so that
the new network resource configuration can be reviewed easily.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Reviewed-by: Chandan Babu R <chandanbabu@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>