Enable EPEL repo after installing epel-release
Not all of the roles that install the epel-release package then go
on to enable it on Oracle Linux, where installing epel-release
leaves the repo disabled, so the repo must then be enabled as a
separate action.

To ensure EPEL is installed and enabled properly in all cases,
de-duplicate the copies of the EPEL installation code into a single
Ansible playbook.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>