terraform/Azure: Update list of available virtual machine sizes
The current set of vmsize options are previous generation, according to

https://learn.microsoft.com/en-us/azure/virtual-machines/sizes/previous-gen-sizes-list#general-purpose-previous-gen-sizes

and are thus capacity-limited. Add a few newer generation options.
in the x86_64 family.

It wouldn't be hard to add ARM systems as well, but that needs boot
image options that can select ARM-based boot images, which is a little
more work than I want to tackle here.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>