terraform/Azure: Do not create a storage account
Recently, "make bringup" started timing out while creating the
storage account resource. I can't get rid of the stuck storage
account for hours, even via the Azure console.

It appears that this is a common enough problem that there is
official Microsoft Azure documentation describing the problem and
suggesting how to get the account unstuck and then deleted.

Needless to say, this is no good for automation. But it looks like
the storage account is needed only for the boot_diagnostics argument
of an azurerm_linux_virtual_machine resource, and that argument is
optional. We can remove it for now -- if it proves to be necessary,
a persistent storage account can be created on the Azure console,
and its resource ID specified via Kconfig.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>