ltp: add makefile target to show test results
Under the hood, it more or less just does 'find ... | xargs cat'.

By default, the "*.log" files will be shown for the most recent kernel
run.  You can show the results for a different kernel by overriding the
LAST_KERNEL variable, e.g.

$ LAST_KERNEL=6.11.9-100.fc39.x86_64 make ltp-show-results

You can change the files being shown by overriding the PATTERN variable.
For example, to just see the failures:

$ PATTERN="-name \"*.failed\"" make ltp-show-results

Note that since the ltp log files have a date and timestamp in the
filename, it's possible to accumulate logs from multiple test runs.  If
you're showing the results from the most recent kernel, then we'll only
show the logs from the most recent run (i.e. logs that are in the
last-run directory). But if you're showing the results from an older
kernel, then we'll show all the logs in that directory (but you can
override that too if you get creative with the PATTERN variable).

Signed-off-by: Scott Mayhew <smayhew@redhat.com>