terraform/OCI: Create a set of multiple generic block devices
When provisioning on OCI, terraform creates one block device for
the /data file system, and one for sparse files. This is unlike
other provisioning methods (guestfs and AWS being the primary
examples) which instead create a set of generic block devices and
then set up the sparse files or /data file system on one of those.

Luis and Chandan have agreed to changing OCI to work like the
other terraform providers and guestfs.

Reviewed-by: Luis Chamberlain <mcgrof@kernel.org>
Reviewed-by: Chandan Babu R <chandanbabu@kernel.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>