Revert "bringup: devconfig: fix python3-apt debian issue"
This reverts commit bcea53d0710cc0722520430277c80279e90e0863

Reverting fixes [1] but we need to ensure APT packages are always
uptodate:

Commit bcea53d0710cc0722520430277c80279e90e0863 ("bringup: devconfig:
fix python3-apt debian issue") was originally added as a workaround for
a Debian python3-apt problem. However, in the Debian bug report thread
[2][3] was revealed that the real problem was:

"APT has a fluctuating ABI that
depends on which std:: symbols it calls, and which of those calls
the compiler inlines and omits the symbols for, so it can change
with every update to APT or the compiler."

To address the problem, we ensure that the distribution and especially
the apt packages, are always updated to the latest available versions.
This update is performed during bringup, when the image is first booted
(firstboot).

Additionally, a new firstconfig target has been added to upgrade the
system even if it has already booted with an outdated image.

[1]
https://github.com/linux-kdevops/linux-mm-kpd/actions/runs/13677761684/job/38242628757

[2]
https://lore.kernel.org/kdevops/20250208134034.GA4079710@debian.org/T/#u

[3]
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1095111

Reported-by: Luis Chamberlain <mcgrof@kernel.org>
Acked-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Daniel Gomez <da.gomez@samsung.com>