linux-mirror: Fix installation on Fedora systems
I noticed that git-daemon was not installed when I set up the git
mirror on my Fedora 40 control host.

If "include_tasks:" has a tag, the included tasks also need to have a
"tags:" attribute or they will not be run.

Also, because my lab is already behind a firewall gateway, there
happens to be no firewall running on my controller, so the "enable
the git port" step fails with:

firewall is not currently running, unable to perform immediate actions without a running firewall daemon

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>