linux-mirror: Ensure git-daemon runs as the Ansible user
In my initial installation, git-daemon emits the "safe.directory"
warning and exits; remote attempts to access the repos there fail.

git-daemon needs to run as the same user that owns the repos to
avoid this warning. linux-mirror doesn't currently have a consistent
"user" for this: the instructions say "create a /mirror directory"
but then later tasks in linux-mirror try to create /mirror owned
by root.

This patch doesn't address that larger problem, but it simply
adds a setting under the daemon's [Service] stanza to make it
run as the user that is setting up the mirror.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>