codereadyrepo: Avoid running terraform tasks when they are unneeded
Scott Mayhew notes that the last two steps in the codereadyrepo
role also need to be gated behind the "is there a custom yum repo?"
check. Otherwise the play fails when there is a custom repo on
a non-Fedora Red Hat distribution.

Let's take the opportunity to re-organize this logic a little bit
so that the relevant checking is done just once rather than on
every step.

Reported-by: Scott Mayhew <smayhew@redhat.com>
Fixes: 69bb0dc56ae6 ("codereadyrepo: Refactor to prepare for more complex usage scenarios")
Reviewed-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>